requirements.txt
python-dotenv
pyyaml
llama-index
neo4j
langchain
pyvis
pypdf
Ipython
llama-index-graph-stores-neo4j
llama-index-llms-ollama
llama-index-llms-huggingface
llama-index-embeddings-huggingface
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
bitsandbytes
sentencepiece
accelerate
py2neo
ipywidgets
openai
sentence-transformers
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
